Rework the perform_expect function
The script is compatible with Python 2 and Python 3. However, Python 3
has introduced a difference between text and binary data. Generally, it
means that logged messages may vary slightly depending on the version of
Python. This commit reworks the perform_expect function to make logged
messages independent of the version of Python.

While being there, also wrap some variable parts of logged messages with
single quotes to easily see where they start and end.

Also, move the bytes-to-string conversion outside the strip_special
function. This makes the function more versatile and allows to use the
correct encoding when decoding the OVMF serial console.

Signed-off-by: Jan Hlavac <jhlavac@redhat.com>