Keep output sources in sync
The Python logging module calls the flush() method every time a new
message is logged. Therefore, this should also be done when printing the
output of the QEMU process. Otherwise, these output sources will not be
in sync.

While being there, also make some changes to improve readability of the
output:

 - Append '\n' to the output of the QEMU process if a new message is
   expected to be logged. This ensures that every logged message starts
   on a new line.

 - Append the trailing '\n' to the output of the QEMU process if it is
   the only output source.

Signed-off-by: Jan Hlavac <jhlavac@redhat.com>