Fix and simplify the verbosity level logic
The script supports different levels of verbosity. However, decreasing
or increasing the verbosity by one level does not lead to any change.
This means that a single --quiet or --verbose argument leads to the same
output as when none of them is used. To see a change in verbosity, it
must be decreased or increased by two levels. The described logic does
not make much sense, so I consider it a bug. Fixing this bug also leads
to an overall simplification of the verbosity level logic.

Signed-off-by: Jan Hlavac <jhlavac@redhat.com>